// packages/medusa-react-admin/src/lib/models/AdminPostProductCategoriesCategoryParams.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type AdminPostProductCategoriesCategoryParams = {
  /**
   * (Comma separated) Which fields should be expanded in each product category.
   */
  expand?: string;
  /**
   * (Comma separated) Which fields should be retrieved in each product category.
   */
  fields?: string;
};